# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: MindSafe Mental Health Platform API
  version: 1.0.0
  description: |
    Privacy-first AI mental health platform. All endpoints under `/api` are served by the
    API Gateway (Express) which proxies ML operations to internal Python microservices.

    ## Authentication
    Most endpoints require a Bearer token in the `Authorization` header.
    Obtain tokens via `/api/login` or `/api/auth/google`. Access tokens expire in 15 minutes;
    use `/api/refresh-token` to get a new pair.

    ## Error Format
    All errors return `{ "message": "..." }` with an appropriate HTTP status code.
  contact:
    name: MindSafe Team
  license:
    name: Private
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://mindsafe-api.onrender.com
    description: Free-tier production (Render)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `/api/login`. Expires in 15 minutes.
  schemas:
...
    AnalyzeTextRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          example: "I feel so hopeless and I don't know what to do"
        user_id:
          type: string
        conversation_id:
          type: string
    TextAnalysisResponse:
      type: object
      properties:
        text:
          type: string
        sentiment:
          type: object
          properties:
            sentiment:
              type: string
              enum: [VERY_NEGATIVE, NEGATIVE, NEUTRAL, POSITIVE, VERY_POSITIVE]
            score:
              type: number
            confidence:
              type: number
        emotion:
          type: object
          properties:
            emotion:
              type: string
              enum: [JOY, SADNESS, ANGER, FEAR, SURPRISE, DISGUST, TRUST, ANTICIPATION]
            confidence:
              type: number
            all_emotions:
              type: object
              additionalProperties:
                type: number
        crisis_detection:
          type: object
          properties:
            is_crisis:
              type: boolean
            crisis_level:
              type: string
              enum: [LOW, MEDIUM, HIGH, CRITICAL]
            crisis_keywords:
              type: array
              items:
                type: string
            risk_score:
              type: number
            reason:
              type: string
        language:
          type: string
        processing_time_ms:
          type: number
        timestamp:
          type: string
          format: date-time

    # ── Crisis ──
...
paths:
...
  /analyze:
    post:
      tags: [Emotion Detection]
      summary: Analyze text for emotion, sentiment, and crisis signals
      operationId: analyzeText
      description: |
        **Internal service** (port 8001). Not exposed to clients directly.
        Called by the chatbot service and API gateway internally.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeTextRequest"
      responses:
        "200":
          description: Full text analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TextAnalysisResponse"
              example:
                text: "I feel so hopeless and I don't know what to do"
                sentiment:
                  sentiment: VERY_NEGATIVE
                  score: -0.89
                  confidence: 0.94
                emotion:
                  emotion: SADNESS
                  confidence: 0.87
                  all_emotions:
                    SADNESS: 0.87
                    FEAR: 0.45
                    ANGER: 0.12
                crisis_detection:
                  is_crisis: true
                  crisis_level: HIGH
                  crisis_keywords: ["hopeless"]
                  risk_score: 0.78
                  reason: "Keyword 'hopeless' detected with negative sentiment"
                processing_time_ms: 42.5
  /batch-analyze:
    post:
      tags: [Emotion Detection]
      summary: Batch analyze multiple texts
      operationId: batchAnalyze
      description: "**Internal service** (port 8001)."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [texts]
              properties:
                texts:
                  type: array
                  items:
                    type: string
                user_id:
                  type: string
      responses:
        "200":
          description: Array of analysis results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/TextAnalysisResponse"

  # ── Crisis Detection (internal :8003) ──
...